requirements.txt
davey==0.1.4
discord.py==2.7.1
ffmpeg==1.4
finance-datareader==0.9.102
PyNaCl==1.5.0
python-dotenv==1.2.2
yt-dlp==2026.3.3
google-genai==1.68.0
